Implement business logic for `allocation-requests`. Long -> Integer (ref. [SQLite primary key](https://www.sqlite.org/datatypes.html))

+++ src/main/java/io/github/starrybleu/sideproject0/api/TakingAllocationRequestReqBody.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class TakingAllocationRequestReqBody {
-    Long driverNo;
+    Integer driverNo;
 }
